Add obstacles and their movement

+++ src/main/java/ObstacleSpawn.java
@@ -3,10 +3,10 @@ package main.java;
 public class ObstacleSpawn {
     private String name;
     private double speed;
-    private double y;
+    private int y;
     private double delta;
 
-    public ObstacleSpawn(String name, double speed, double y, double delta) {
+    public ObstacleSpawn(String name, double speed, int y, double delta) {
         this.name = name;
         this.speed = speed;
         this.y = y;
@@ -15,6 +15,6 @@ public class ObstacleSpawn {
 
     public String getName() { return this.name; }
     public double getSpeed() { return this.speed; }
-    public double getY() { return this.y; }
+    public int getY() { return this.y; }
     public double getDelta() { return this.delta; }
 }
